server/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
clerk_django==1.0.3
cryptography==42.0.5
dj-database-url==2.3.0
Django==6.0
django-appconf==1.2.0
django-cors-headers==4.9.0
django-cryptography==1.1
idna==3.11
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
requests==2.31.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
djangorestframework
vercel-blob
